Remove legacy track + playlist indexing code
