+++ discovery-provider/integration_tests/tasks/test_index_operations.py
@@ -373,76 +373,6 @@ def test_index_operations_tx_receipts_fetch_error(
         assert error["count"] == 1
 
 
-def test_index_operations_tx_parse_error(celery_app, celery_app_contracts, mocker):
-    """
-    Confirm indexer throws IndexingError when the parser throws an error
-    """
-    task = celery_app.celery.tasks["update_discovery_provider"]
-    db = task.db
-    web3 = celery_app_contracts["web3"]
-
-    mocker.patch(
-        "src.utils.helpers.get_final_poa_block",
-        return_value=final_poa_block,
-        autospec=True,
-    )
-
-    # patch parse track event to raise an exception
-    def parse_track_event(*_):
-        raise Exception("Broken parser")
-
-    mocker.patch(
-        "src.tasks.tracks.parse_track_event",
-        side_effect=parse_track_event,
-        autospec=True,
-    )
-
-    seed_data = seed_contract_data(task, celery_app_contracts, web3)
-    mocker.patch(
-        "src.utils.cid_metadata_client.CIDMetadataClient._get_gateway_endpoints",
-        return_value=["https://test-content-node.audius.co"],
-        autospec=True,
-    )
-
-    mock_response = MockResponse(seed_data["track_metadata"], 200)
-    mocker.patch(
-        "aiohttp.ClientSession.get",
-        return_value=mock_response,
-        autospec=True,
-    )
-    current_block = None
-    latest_block = None
-    try:
-        with db.scoped_session() as session:
-            # Catch up the indexer
-            current_block_query = session.query(Block).filter_by(is_current=True).all()
-            current_block = (
-                current_block_query[0].number if len(current_block_query) > 0 else 0
-            )
-            latest_block = web3.eth.getBlock("latest", True)
-            while current_block < latest_block.number:
-                # Process a bunch of blocks to make sure we covered everything
-                task.run()
-                current_block_query = (
-                    session.query(Block).filter_by(is_current=True).all()
-                )
-                current_block = (
-                    current_block_query[0].number if len(current_block_query) > 0 else 0
-                )
-        assert False
-    except IndexingError:
-        error = get_indexing_error(redis)
-        errored_block_in_db_results = (
-            session.query(Block).filter_by(number=error["blocknumber"]).all()
-        )  # should not exist
-        errored_block_in_db = len(errored_block_in_db_results) != 0
-        # when errored block is in db, it breaks the consensus mechanism
-        # for discovery nodes staying in sync
-        assert not errored_block_in_db
-        assert error["message"] == "Broken parser"
-        assert error["count"] == 1
-
-
 def test_index_operations_indexing_error_on_commit(
     celery_app, celery_app_contracts, monkeypatch, mocker
 ):
