+++ discovery-provider/src/tasks/entity_manager/playlist.py
@@ -6,6 +6,7 @@ from typing import Dict, Set
 from src.challenges.challenge_event import ChallengeEvent
 from src.challenges.challenge_event_bus import ChallengeEventBus
 from src.models.playlists.playlist import Playlist
+from src.models.playlists.playlist_route import PlaylistRoute
 from src.tasks.entity_manager.utils import (
     PLAYLIST_ID_OFFSET,
     Action,
@@ -13,9 +14,89 @@ from src.tasks.entity_manager.utils import (
     ManageEntityParameters,
     copy_record,
 )
-from src.tasks.playlists import update_playlist_routes_table
+from src.tasks.task_helpers import generate_slug_and_collision_id
+from src.utils import helpers
 
 logger = logging.getLogger(__name__)
+
+
+def update_playlist_routes_table(session, playlist_record, pending_playlist_routes):
+    logger.info(
+        f"index.py | playlists.py | Updating playlist routes for {playlist_record.playlist_id}"
+    )
+    # Get the title slug, and set the new slug to that
+    # (will check for conflicts later)
+    new_playlist_slug_title = helpers.sanitize_slug(
+        playlist_record.playlist_name, playlist_record.playlist_id
+    )
+    new_playlist_slug = new_playlist_slug_title
+
+    # Find the current route for the playlist
+    # Check the pending playlist route updates first
+    prev_playlist_route_record = next(
+        (
+            route
+            for route in pending_playlist_routes
+            if route.is_current and route.playlist_id == playlist_record.playlist_id
+        ),
+        None,
+    )
+
+    # Then query the DB if necessary
+    if prev_playlist_route_record is None:
+        prev_playlist_route_record = (
+            session.query(PlaylistRoute)
+            .filter(
+                PlaylistRoute.playlist_id == playlist_record.playlist_id,
+                PlaylistRoute.is_current == True,
+            )  # noqa: E712
+            .one_or_none()
+        )
+
+    if prev_playlist_route_record:
+        if prev_playlist_route_record.title_slug == new_playlist_slug_title:
+            # If the title slug hasn't changed, we have no work to do
+            logger.info(f"not changing for {playlist_record.playlist_id}")
+            return
+        # The new route will be current
+        prev_playlist_route_record.is_current = False
+
+    new_playlist_slug, new_collision_id = generate_slug_and_collision_id(
+        session,
+        PlaylistRoute,
+        playlist_record.playlist_id,
+        playlist_record.playlist_name,
+        playlist_record.playlist_owner_id,
+        pending_playlist_routes,
+        new_playlist_slug_title,
+        new_playlist_slug,
+    )
+
+    # Add the new playlist route
+    new_playlist_route = PlaylistRoute()
+    new_playlist_route.slug = new_playlist_slug
+    new_playlist_route.title_slug = new_playlist_slug_title
+    new_playlist_route.collision_id = new_collision_id
+    new_playlist_route.owner_id = playlist_record.playlist_owner_id
+    new_playlist_route.playlist_id = playlist_record.playlist_id
+    new_playlist_route.is_current = True
+    new_playlist_route.blockhash = playlist_record.blockhash
+    new_playlist_route.blocknumber = playlist_record.blocknumber
+    new_playlist_route.txhash = playlist_record.txhash
+    session.add(new_playlist_route)
+
+    # Add to pending playlist routes so we don't add the same route twice
+    pending_playlist_routes.append(new_playlist_route)
+
+    logger.info(
+        f"index.py | playlists.py | Updated playlist routes for {playlist_record.playlist_id} with slug {new_playlist_slug} and owner_id {new_playlist_route.owner_id}"
+    )
+
+
+def get_playlist_events_tx(update_task, event_type, tx_receipt):
+    return getattr(update_task.playlist_contract.events, event_type)().processReceipt(
+        tx_receipt
+    )
 
 
 def validate_playlist_tx(params: ManageEntityParameters):
